Add wechatmp_service channel
